Item entity, business logic
