+++ src/main/java/jpabook/jpashop/domain/OrderItem.java
@@ -1,6 +1,6 @@
 package jpabook.jpashop.domain;
 
-import jpabook.jpashop.domain.item.Item;
+import jpabook.jpashop.item.Item;
 import lombok.Getter;
 import lombok.Setter;
