+++ src/main/java/jpabook/jpashop/exception/NotEnoughStockException.java
@@ -1,0 +1,21 @@
+package jpabook.jpashop.exception;
+
+public class NotEnoughStockException extends RuntimeException {
+
+    public NotEnoughStockException() {
+        super();
+    }
+
+    public NotEnoughStockException(String message) {
+        super(message);
+    }
+
+    public NotEnoughStockException(String message, Throwable cause) {
+        super(message, cause);
+    }
+
+    public NotEnoughStockException(Throwable cause) {
+        super(cause);
+    }
+
+}
