+++ src/main/java/jpabook/jpashop/item/Album.java
@@ -1,4 +1,4 @@
-package jpabook.jpashop.domain.item;
+package jpabook.jpashop.item;
 
 import lombok.Getter;
 import lombok.Setter;
@@ -7,10 +7,7 @@ import javax.persistence.DiscriminatorValue;
 import javax.persistence.Entity;
 
 @Entity
-@DiscriminatorValue("B")
+@DiscriminatorValue("A")
 @Getter @Setter
-public class Book extends Item {
-
-    private String author;
-    private String isbn;
+public class Album extends Item {
 }
