+++ src/main/java/jpabook/jpashop/item/Book.java
@@ -1,4 +1,4 @@
-package jpabook.jpashop.domain.item;
+package jpabook.jpashop.item;
 
 import lombok.Getter;
 import lombok.Setter;
@@ -7,7 +7,10 @@ import javax.persistence.DiscriminatorValue;
 import javax.persistence.Entity;
 
 @Entity
-@DiscriminatorValue("A")
+@DiscriminatorValue("B")
 @Getter @Setter
-public class Album extends Item {
+public class Book extends Item {
+
+    private String author;
+    private String isbn;
 }
