+++ src/main/java/jpabook/jpashop/item/Item.java
@@ -1,6 +1,7 @@
-package jpabook.jpashop.domain.item;
+package jpabook.jpashop.item;
 
 import jpabook.jpashop.domain.Category;
+import jpabook.jpashop.exception.NotEnoughStockException;
 import lombok.Getter;
 import lombok.Setter;
 
@@ -24,4 +25,22 @@ public abstract class Item {
 
     @ManyToMany(mappedBy = "items")
     private List<Category> categories = new ArrayList<>();
+
+    //== 비지니스 로직 ==//
+    /*
+        stock 증가
+    * */
+    public void addStock(int quantity){
+        this.stockQuantity += quantity;
+    }
+    /*
+        stock 감소
+    * */
+    public void removeStock(int quantity){
+        int restStock = this.stockQuantity - quantity;
+        if (restStock <0){
+            throw new NotEnoughStockException("need more stock");
+        }
+        this.stockQuantity = restStock;
+    }
 }
